Progress: Game all working, need fix compare user prediction

+++ src/assignment1/Athlete.java
@@ -6,6 +6,6 @@ abstract public class Athlete extends Participant {
 		super(id, name, age, state, type);
 	}
 
-	abstract public double compete();
+	abstract public int compete();
 
 }
